Don't update search results every keystroke SL #2734
If the log file is long and we search for new results every time that a key is
pressed this action could consume a lot of CPU. Even more if we search for a
single char.

So, I changed the way that it works to update the search results every time the
user presses the Enter key or the _AUTOSEARCH_TIMEOUT expired like the Sugar's Home does.

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>
Reviewed-by: James Cameron <quozl@laptop.org>